third_party/blink/web_tests/platform/generic/fast/forms/textarea/cols-attribute-expected.txt

HTMLTextAreaElement cols attribute test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.createElement("textarea").cols is 20
PASS colsAttributeEffect("") is 20
PASS colsAttributeEffect("1") is 1
PASS colsAttributeEffect("2") is 2
PASS colsAttributeEffect("10") is 10
PASS colsAttributeEffect("0") is 20
PASS colsAttributeEffect("-1") is 20
PASS colsAttributeEffect("1x") is 1
PASS colsAttributeEffect("1.") is 1
PASS colsAttributeEffect("1.9") is 1
PASS colsAttributeEffect("2x") is 2
PASS colsAttributeEffect("2.") is 2
PASS colsAttributeEffect("2.9") is 2
PASS colsAttributeEffect("a") is 20
PASS colsAttributeEffect("7") is 20
PASS colsAttributeEffect("  7") is 7
PASS colsAttributeEffect(arabicIndicDigitOne) is 20
PASS colsAttributeEffect("2" + arabicIndicDigitOne) is 2
PASS colsAttributeEffect("2147483646") is 2147483646
PASS colsAttributeEffect("2147483647") is 2147483647
PASS colsAttributeEffect("2147483648") is 20
PASS colsAttributeEffect("2147483649") is 20
PASS colsAttributeEffect("4294967295") is 20
PASS colsAttributeEffect("4294967296") is 20
PASS colsAttributeEffect("4294967297") is 20
PASS successfullyParsed is true

TEST COMPLETE
